use rubocop

+++ lib/epdiff.rb
@@ -1,4 +1,6 @@
-require "epdiff/version"
+# frozen_string_literal: true
+
+require 'epdiff/version'
 require 'optparse'
 require 'tmpdir'
 require 'fileutils'
@@ -6,9 +8,9 @@ require 'zip'
 require 'pastel'
 require 'tty/file'
 
-
+# EPUB diff class
 class Epdiff
-  TEXT_EXT = %w(xhtml html xml opf css txt)
+  TEXT_EXT = %w[xhtml html xml opf css txt].freeze
 
   def self.execute(*args)
     new.execute(*args)
@@ -24,22 +26,22 @@ class Epdiff
 
   def unzip(filename, dir)
     Zip::InputStream.open(filename) do |zio|
-      while entry = zio.get_next_entry
-        if  entry.file?
-          entry_filename = File.join(dir, entry.name)
-          FileUtils.mkdir_p File.dirname(entry_filename)
-          File.binwrite(entry_filename, zio.read)
-        end
+      while entry = zio.get_next_entry # rubocop:disable Lint/AssignmentInCondition
+        next unless entry.file?
+
+        entry_filename = File.join(dir, entry.name)
+        FileUtils.mkdir_p File.dirname(entry_filename)
+        File.binwrite(entry_filename, zio.read)
       end
     end
   end
 
-  def show_diff(file1, file2, work_dir)
+  def show_diff(_file1, _file2, work_dir)
     Dir.chdir(work_dir) do
-      files1 = Dir.glob("file1/**/*")
-      files2 = Dir.glob("file2/**/*")
-      files1_s = files1.map{|d| d.sub(%r{^file1/}, '')}
-      files2_s = files2.map{|d| d.sub(%r{^file2/}, '')}
+      files1 = Dir.glob('file1/**/*')
+      files2 = Dir.glob('file2/**/*')
+      files1_s = files1.map { |d| d.sub(%r{^file1/}, '') }
+      files2_s = files2.map { |d| d.sub(%r{^file2/}, '') }
       files_common = files1_s & files2_s
       files_common.each do |path|
         if File.file?("file1/#{path}")
@@ -65,39 +67,40 @@ class Epdiff
 
   def text_file?(path)
     extname = File.extname(path).sub(/\A\./, '')
-    if TEXT_EXT.include?(extname)
-      extname
-    else
-      nil
-    end
+    extname if TEXT_EXT.include?(extname)
   end
 
   def text_diff(_path, path1, path2)
-    diff = @color ? TTY::File.diff(path1, path2, verbose: false) : TTY::File.diff(path1, path2, verbose: false, color: nil)
-    if diff != "No differences found\n" && diff.strip != ""
-      @exit_code = 1
-      print diff
-    end
+    diff = if @color
+             TTY::File.diff(path1, path2, verbose: false)
+           else
+             TTY::File.diff(path1, path2, verbose: false, color: nil)
+           end
+
+    return if diff == "No differences found\n" || diff.strip == ''
+
+    @exit_code = 1
+    print diff
   end
 
   def binary_diff(path, path1, path2)
     content1 = File.binread(path1)
     content2 = File.binread(path2)
-    if content1 != content2
-      @exit_code = 1
-      message = "DIFF: #{path} has some differences.\n"
-      print @color ? @cyan.call(message) : message
-    end
+
+    return if content1 == content2
+
+    @exit_code = 1
+    message = "DIFF: #{path} has some differences.\n"
+    print @color ? @cyan.call(message) : message
   end
 
   def execute(*args)
-
     tmpdir = work_dir = nil
 
-    opts = OptionParser.new do |opts|
+    options = OptionParser.new do |opts|
       opts.banner = "Usage: epdiff [options] [filename] [filename]\n"
 
-      opts.on('-t','--tmpdir DIR', 'Set tepmorary directory') do |dir|
+      opts.on('-t', '--tmpdir DIR', 'Set tepmorary directory') do |dir|
         tmpdir = dir
       end
 
@@ -117,20 +120,19 @@ class Epdiff
     end
 
     begin
-
-      opts.parse!(args)
+      options.parse!(args)
 
       if args.size != 2
         # invalid option
-        puts opts
+        puts options
         exit
       end
 
       work_dir = tmpdir || Dir.mktmpdir
 
       file1, file2 = *args
-      FileUtils.mkdir_p(work_dir+"/file1")
-      FileUtils.mkdir_p(work_dir+"/file2")
+      FileUtils.mkdir_p("#{work_dir}/file1")
+      FileUtils.mkdir_p("#{work_dir}/file2")
 
       unzip(file1, "#{work_dir}/file1")
       unzip(file2, "#{work_dir}/file2")
@@ -139,14 +141,10 @@ class Epdiff
       show_diff(file1, file2, work_dir)
 
       @exit_code
-    rescue => e
+    rescue StandardError => e
       warn e
-      puts opts
     ensure
-      if work_dir && !tmpdir
-        FileUtils.rm_rf(work_dir)
-      end
+      FileUtils.rm_rf(work_dir) if work_dir && !tmpdir
     end
   end
-
 end
